+++ Test/UnitTests/TestExamples/TestExample2RoleChangedDetectorService.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) 2022 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
 // Licensed under MIT license. See License.txt in the project root for license information.
 
-using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using Example2.WebApiWithToken.IndividualAccounts.ClaimsChangeCode;
@@ -20,24 +19,24 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Test.UnitTests.TestExamples;
 
-public class TestExample2RoleChangeDetectorService
+public class TestExample2RoleChangedDetectorService
 {
     private readonly ITestOutputHelper _output;
 
-    public TestExample2RoleChangeDetectorService(ITestOutputHelper output)
+    public TestExample2RoleChangedDetectorService(ITestOutputHelper output)
     {
         _output = output;
     }
 
 
     [Fact]
-    public async Task TestRoleChangeDetectorService_CheckUserAndRoles()
+    public async Task TestRoleChangedDetectorService_CheckUserAndRoles()
     {
         //SETUP
         var authOptions = new AuthPermissionsOptions { InternalData = { EnumPermissionsType = typeof(TestEnum) } };
         var options = SqliteInMemory.CreateOptions<AuthPermissionsDbContext>();
         var stubFsCache = new StubFileStoreCacheClass();
-        var context = new AuthPermissionsDbContext(options, new List<IDatabaseStateChangeEvent> { new RoleChangeDetectorService(stubFsCache, authOptions)});
+        var context = new AuthPermissionsDbContext(options, new List<IDatabaseStateChangeEvent> { new RoleChangedDetectorService(stubFsCache, authOptions)});
         context.Database.EnsureCreated();
 
         await context.SetupRolesInDbAsync();
@@ -61,13 +60,13 @@ public class TestExample2RoleChangeDetectorService
     }
 
     [Fact]
-    public async Task TestRoleChangeDetectorService_AddUserToRole()
+    public async Task TestRoleChangedDetectorService_AddUserToRole()
     {
         //SETUP
         var authOptions = new AuthPermissionsOptions { InternalData = { EnumPermissionsType = typeof(TestEnum) } };
         var options = SqliteInMemory.CreateOptions<AuthPermissionsDbContext>();
         var stubFsCache = new StubFileStoreCacheClass();
-                var context = new AuthPermissionsDbContext(options, new List<IDatabaseStateChangeEvent> { new RoleChangeDetectorService(stubFsCache, authOptions)});
+                var context = new AuthPermissionsDbContext(options, new List<IDatabaseStateChangeEvent> { new RoleChangedDetectorService(stubFsCache, authOptions)});
         context.Database.EnsureCreated();
 
         await context.SetupRolesInDbAsync();
@@ -89,16 +88,18 @@ public class TestExample2RoleChangeDetectorService
         }
         var allChanges = stubFsCache.GetAllKeyValues();
         allChanges.Count.ShouldEqual(1);
+        allChanges.First().Key.ShouldEqual("ReplacementPermissionsUser1");
+        allChanges.First().Value.Select(x => (int)x).ShouldEqual(new []{1,2});
     }
 
     [Fact]
-    public async Task TestRoleChangeDetectorService_RemoveUserToRole()
+    public async Task TestRoleChangedDetectorService_RemoveUserToRole()
     {
         //SETUP
         var authOptions = new AuthPermissionsOptions { InternalData = { EnumPermissionsType = typeof(TestEnum) } };
         var options = SqliteInMemory.CreateOptions<AuthPermissionsDbContext>();
         var stubFsCache = new StubFileStoreCacheClass();
-                var context = new AuthPermissionsDbContext(options, new List<IDatabaseStateChangeEvent> { new RoleChangeDetectorService(stubFsCache, authOptions)});
+                var context = new AuthPermissionsDbContext(options, new List<IDatabaseStateChangeEvent> { new RoleChangedDetectorService(stubFsCache, authOptions)});
         context.Database.EnsureCreated();
 
         await context.SetupRolesInDbAsync();
@@ -111,7 +112,7 @@ public class TestExample2RoleChangeDetectorService
             new StubSyncAuthenticationUsersFactory(), authOptions);
 
         //ATTEMPT
-        await authAdmin.UpdateUserAsync("User2", roleNames: new List<string> { "Role2" });
+        await authAdmin.UpdateUserAsync("User3", roleNames: new List<string> { "Role1" });
 
         //VERIFY
         foreach (var user in authAdmin.QueryAuthUsers().Include(x => x.UserRoles).ToList())
@@ -120,16 +121,51 @@ public class TestExample2RoleChangeDetectorService
         }
         var allChanges = stubFsCache.GetAllKeyValues();
         allChanges.Count.ShouldEqual(1);
+        allChanges.First().Key.ShouldEqual("ReplacementPermissionsUser3");
+        allChanges.First().Value.Select(x => (int)x).ShouldEqual(new[] { 1 });
     }
 
     [Fact]
-    public async Task TestRoleChangeDetectorService_ChangeRoleToPermissions()
+    public async Task TestRoleChangedDetectorService_NoRoles()
     {
         //SETUP
         var authOptions = new AuthPermissionsOptions { InternalData = { EnumPermissionsType = typeof(TestEnum) } };
         var options = SqliteInMemory.CreateOptions<AuthPermissionsDbContext>();
         var stubFsCache = new StubFileStoreCacheClass();
-                var context = new AuthPermissionsDbContext(options, new List<IDatabaseStateChangeEvent> { new RoleChangeDetectorService(stubFsCache, authOptions)});
+        var context = new AuthPermissionsDbContext(options, new List<IDatabaseStateChangeEvent> { new RoleChangedDetectorService(stubFsCache, authOptions) });
+        context.Database.EnsureCreated();
+
+        await context.SetupRolesInDbAsync();
+        context.AddMultipleUsersWithRolesInDb();
+
+        context.ChangeTracker.Clear();
+        stubFsCache.ClearAll();
+
+        var authAdmin = new AuthUsersAdminService(context,
+            new StubSyncAuthenticationUsersFactory(), authOptions);
+
+        //ATTEMPT
+        await authAdmin.UpdateUserAsync("User2", roleNames: new List<string> { });
+
+        //VERIFY
+        foreach (var user in authAdmin.QueryAuthUsers().Include(x => x.UserRoles).ToList())
+        {
+            _output.WriteLine($"UserId = {user.UserId}, Roles = {string.Join(", ", user.UserRoles.Select(x => x.RoleName))}");
+        }
+        var allChanges = stubFsCache.GetAllKeyValues();
+        allChanges.Count.ShouldEqual(1);
+        allChanges.First().Key.ShouldEqual("ReplacementPermissionsUser2");
+        allChanges.First().Value.ShouldEqual("");
+    }
+
+    [Fact]
+    public async Task TestRoleChangedDetectorService_ChangeRoleToPermissions()
+    {
+        //SETUP
+        var authOptions = new AuthPermissionsOptions { InternalData = { EnumPermissionsType = typeof(TestEnum) } };
+        var options = SqliteInMemory.CreateOptions<AuthPermissionsDbContext>();
+        var stubFsCache = new StubFileStoreCacheClass();
+                var context = new AuthPermissionsDbContext(options, new List<IDatabaseStateChangeEvent> { new RoleChangedDetectorService(stubFsCache, authOptions)});
         context.Database.EnsureCreated();
 
         await context.SetupRolesInDbAsync();
@@ -151,5 +187,7 @@ public class TestExample2RoleChangeDetectorService
         var allChanges = stubFsCache.GetAllKeyValues();
         allChanges.Count.ShouldEqual(2);
         allChanges.Keys.ShouldEqual(new []{ "ReplacementPermissionsUser2", "ReplacementPermissionsUser3" });
+        allChanges["ReplacementPermissionsUser2"].Select(x => (int)x).ShouldEqual(new[] { 1, 3 });
+        allChanges["ReplacementPermissionsUser3"].Select(x => (int)x).ShouldEqual(new[] { 1, 3 });
     }
 }