+++ Test/UnitTests/TestExamples/UnitTestExample.cs
@@ -49,7 +49,8 @@ namespace Test.UnitTests.TestExamples
             var rereadUser = status.Result;
             rereadUser.Email.ShouldEqual(userId.ToLower());
             rereadUser.UserName.ShouldEqual(userId);
-            rereadUser.UserRoles.Select(x => x.RoleName).ShouldEqual(new List<string> { "Tenant Admin", "Area Manager" });
+            rereadUser.UserRoles.OrderBy(x => x.RoleName)
+                .Select(x => x.RoleName).ShouldEqual(new List<string> { "Area Manager", "Tenant Admin" });
             rereadUser.UserTenant.TenantFullName.ShouldEqual("4U Inc.");
         }