+++ ReleaseNotes.md
@@ -2,8 +2,8 @@
 
 ## 3.5.0
 
-- Bug Fix: When using JWT Token authentication the ClaimsCalculator has to set up the nameidentifier
 - Improved feature: AuthPermissionsDbContext now takes mutiple IDatabaseStateChangeEvent
+- Improved feature: No AuthP database event change listeners will be triggered during bulk loading
 - Improved feature: Updated FileStore distributed cache to version ???
 
 ## 3.4.0